test contains  VS2019 solution
