--- .travis.yml
@@ -1,20 +1,0 @@
-sudo: false
-
-language: cpp
-
-compiler:
-  - gcc
-  - clang
-
-addons:
-  apt:
-    sources:
-    - ubuntu-toolchain-r-test
-    packages:
-    - g++
-    - clang
-
-script:
-  - cmake -DCMAKE_BUILD_TYPE=Release
-  - make
-  - ctest
