--- default.nix
@@ -1,10 +1,0 @@
-{
-  nixpkgs ? <nixpkgs>,
-  pkgs ? import nixpkgs {}
-}:
-
-pkgs.stdenv.mkDerivation {
-  name = "pprintpp";
-  src = ./.;
-  nativeBuildInputs = [ pkgs.cmake ];
-}
