+++ test/pprintpp.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29418.71
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "pprintpp", "pprintpp.vcxproj", "{4C73F7F4-53A7-44E8-91A2-C649129B8A18}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4C73F7F4-53A7-44E8-91A2-C649129B8A18}.Debug|x64.ActiveCfg = Debug|x64
+		{4C73F7F4-53A7-44E8-91A2-C649129B8A18}.Debug|x64.Build.0 = Debug|x64
+		{4C73F7F4-53A7-44E8-91A2-C649129B8A18}.Debug|x86.ActiveCfg = Debug|Win32
+		{4C73F7F4-53A7-44E8-91A2-C649129B8A18}.Debug|x86.Build.0 = Debug|Win32
+		{4C73F7F4-53A7-44E8-91A2-C649129B8A18}.Release|x64.ActiveCfg = Release|x64
+		{4C73F7F4-53A7-44E8-91A2-C649129B8A18}.Release|x64.Build.0 = Release|x64
+		{4C73F7F4-53A7-44E8-91A2-C649129B8A18}.Release|x86.ActiveCfg = Release|Win32
+		{4C73F7F4-53A7-44E8-91A2-C649129B8A18}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {301B9B01-7CE8-499B-A238-F91129BBDAF7}
+	EndGlobalSection
+EndGlobal
