+++ test/readme.md
@@ -1,0 +1,17 @@
+
+> Original project is https://github.com/tfc/pprintpp
+> 
+> Please respect the IP of the original project.
+
+# PPRINTPP port to MSVC 
+
+Releases
+
+| Major | Minor | Patch | Comment
+|---|---|---|---|
+|0|1|0| Works with CL 19.23.28106.4 for x86. By default does not depend on std lib. Can not compile larger formating
+
+---
+Original work in this repository is &copy; dbj@dbj.org (dusanjovanovic.org)
+
+Licenese for this repository is **[CC BY SA 4.0](https://creativecommons.org/licenses/by-sa/4.0/)**
